Refactor CLI test mixin
